app/requirements.txt
pytz==2024.1
pymsteams==0.2.3
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
SQLAlchemy==2.0.31
xmltodict==0.13.0